// src/types/rabbitmq.ts
export interface RabbitMQQueueInfo {
  readonly name: string;
  readonly messages: number;
  readonly consumers: number;
  readonly state: string;
  readonly durable: boolean;
  readonly autoDelete: boolean;
  readonly type: string;
}

export interface RabbitMQMessage {
  readonly payload: string;
  readonly payloadEncoding: string;
  readonly exchange: string;
  readonly routingKey: string;
  readonly redelivered: boolean;
  readonly properties: RabbitMQMessageProperties;
}

export interface RabbitMQMessageProperties {
  readonly contentType: string | null;
  readonly contentEncoding: string | null;
  readonly deliveryMode: number | null;
  readonly headers: Record<string, unknown>;
  readonly timestamp: number | null;
  readonly messageId: string | null;
  readonly appId: string | null;
}